Use v2 cloud API
