+++ src/utils/grpc.ts
@@ -1,5 +1,5 @@
 import {createChannel, createClient, Metadata} from 'nice-grpc'
-import {CloudServiceClient, CloudServiceDefinition} from '../proto/depot/cloud/v1/cloud'
+import {CloudServiceClient, CloudServiceDefinition} from '../proto/depot/cloud/v2/cloud'
 import {CLOUD_AGENT_API_URL, CLOUD_AGENT_CONNECTION_TOKEN, CLOUD_AGENT_TF_MODULE_VERSION} from './env'
 
 const channel = createChannel(CLOUD_AGENT_API_URL)
